Move buildReductionStrategyBlockDistribution from gGPUto tCommon
